+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Head
 
+- Deprecated `"boolean"` option for `emptyLineBefore` when using property groups in `rule-properties-order`. Use the new `"always"` or `"never"` option, instead.
+- Added: `"always"` and `"never"` option to `rule-properties-order` `emptyLineBefore` when using property groups
 - Added: `no-browser-hacks` rule.
 - Added: `string-no-newline` rule.
 - Added: `unspecified: "bottomAlphabetical"` option for `rule-properties-order`.